Annotations

Term.Annotations.Annotation

Term.Annotations.Annotation

Term.Annotations.Decoration

Term.Annotations.Decoration

Term.Annotations.half

Term.Annotations.join_interleaved

Term.Annotations.make_spaces

Term.Annotations.overlay_decorations

Term.Annotations.Annotation — Type

struct  Annotation <: AbstractRenderable
    segments::Vector
    measure::Measure
end

Represents a bit of text with some additional annotations. Annotations are additional messages that get printed below the main piece of text.

Example

Annotation("This is the text", "text"=>"this is an annotation")
````
gives

This is the text ──┬─ │ │ ╭─────────────────────────╮ ╰─│ this is an annotation │ ╰─────────────────────────╯

---
This bit:

──┬─ │ │ ╭─────────────────────────╮ ╰─│ this is an annotation │ ╰─────────────────────────╯ ``is called aDecoration`.

A piece of text can have multiple decorations if multiple pares are passed in the function call above.

Term.Annotations.Annotation — Method

Annotation(text::String, annotations::Pair...; kwargs...)

Construct an Annotation.

The argument annotations is a set of Pairs denoting which parts of text gets annotated, with what, and what style information each annotation should have. These pairas can be of the form String=>String if no style information is passed, otherwise String=>(String, String) where the first String in the parentheses is the annotation message and the second the style info. The first String in the Pair should be a substring of text to denote where the annotation occurs.

Term.Annotations.Decoration — Type

struct Decoration nun::Int position::Int underscore::hLine panel::Renderable style::String end

A decoration storing a message to anotate a piece of text. Decorations are used by Annotation to annotate text.

A Decoration looks something like this when rendered.

────┬───                              
    │                                 
    │ ╭─────╮
    ╰─│ MSG │
      ╰─────╯

Term.Annotations.Decoration — Method

Decoration(num::Int, position::Int, message::String, underscore_width::Int, style::String)

Construct a Decoration. When preparing the Panel with the message, the message text gets resized if its too large (based on position and console_width).

Arguments

num number of decoration (for Annotation with multiple decorations)

position position of the decoration start in the text (position of the start of the underscore)

message text going inside the message box

underscore_width: the width of the underscore line.

style: color/style information

Term.Annotations.half — Method

halve and round a number

Term.Annotations.join_interleaved — Method

hstack interleaved elements x ∈ X, y ∈ Y

Term.Annotations.make_spaces — Method

Make a vector of Spacer objects of given widths

Term.Annotations.overlay_decorations — Method

overlay_decorations(decorations::Vector{Decoration})

Overlayed rendering of multiple Decoration object. Given multiple decorations, create a visualization like:

Panel(content; fit=true)                                                                 
──┬── ───┬───  ────┬───                                                                  
  │      │         │                                                                     
  │ ╭─────────────────────────────────╮                                                  
  ╰─│  this is the panel constructor  │                                                  
    ╰─────────────────────────────────╯                                                  
         │         │                                                                     
         │ ╭───────────────────────────────────────────╮                                 
         ╰─│  here you put what goes inside the panel  │                                 
           ╰───────────────────────────────────────────╯                                 
                   │                                                                     
                   │ ╭──────────────────────────────────────────────────────────────────╮
                   ╰─│  Setting this as `true` adjusts the panel's width to fit         │
                     │  `content`. Otherwise `Panel` will have a fixed width            │
                     ╰──────────────────────────────────────────────────────────────────╯

Note: we need all the underscores, the vertical lines to be in the right place and for the Panel with heach message to be over the lines of the following decorations. Spacing should be accounted to make sure all messages are visible.

Note: this function is quite long and nasty, but it seems to work well.