Boxes

Term.Boxes.Box

Term.Boxes.Box

Term.Boxes.BoxLine

Term.Boxes.fit

Term.Boxes.get_lrow

Term.Boxes.get_row

Term.Boxes.get_row

Term.Boxes.get_rrow

Term.Boxes.get_title_row

Term.Boxes.Box — Type

Box

Defines characters to render boxes.

Row names:

┌─┬┐ top │ ││ head ├─┼┤ headrow │ ││ mid ├─┼┤ row ├─┼┤ footrow │ ││ foot └─┴┘ bottom

each row is an instance of BoxLine

Term.Boxes.Box — Method

Box(string)

Construct a Box objet out of a box string.

Term.Boxes.BoxLine — Type

BoxLine

Stores the characters for a line of a Box object.

Term.Boxes.fit — Method

fit(box::Box, widths::Vector{Int})::String

Creates a box.

The box has one of each level type with columns widths specified by a vector of widhts.

Term.Boxes.get_lrow — Method

get_lrow(box::Box, width::Int, level::Symbol)::String

Get a box's row's left part (no righ char)

Get a box's row's right part (no left char) See also get_row, get_rrow.

Term.Boxes.get_row — Method

get_row(box::Box, width::Int, level::Symbol)::String

Get a box's row of given width.

Term.Boxes.get_row — Method

get_row(box, [1, 2, 3], :row)

Gets characters for a row of a Box object.

The level Symbol can be used to specify the box level (:top, :footer...). The total width will be the sum of the widths +2

Term.Boxes.get_rrow — Method

get_rrow(box::Box, width::Int, level::Symbol)::String

Get a box's row's right part (no left char) See also get_row, get_lrow.

Term.Boxes.get_title_row — Method

gettitlerow(row::Symbol, box::Box, title::Union{Nothing, AbstractString}; <keyword arguments>)

Create a box row with a title string.

Can create both titles in the top and bottom row to produce subtitles.

#Arguments:

width::Int: width of line

style::String: style of line

title_style::String: style of title string

justify::Symbol=:left: position of title string

See also get_row.