Color

Term.ANSICode

Term.color.AbstractColor

Term.color.get_color

Term.color.hex2rgb

Term.color.hsl2rgb

Term.color.is_background

Term.color.is_color

Term.color.is_hex_color

Term.color.is_named_color

Term.color.is_rgb_color

Term.ANSICode — Method

ANSICode(color; bg::Bool=false)

Create ANSI tags for colors.

Term.color.AbstractColor — Type

AbstractColor

Abstract color type.

Term.color.get_color — Method

get_color(string::String; bg=false)::AbstractColor

Extract a color type from a string with color information.

Term.color.hex2rgb — Method

hex2rgb(hex::String)

Converts a string hex color code to a RGB color

Term.color.hsl2rgb — Method

Convert HSL values to RGB values, return a markup color string.

Term.color.is_background — Method

is_background(string::String)::Bool

Check if a string represents background color information, of any type.

Term.color.is_color — Method

is_color(string::String)::Bool

Check if a string represents color information, of any type.

Term.color.is_hex_color — Method

is_hex_color(string::String)::Bool

Check if a string represents a hex color.

Term.color.is_named_color — Method

is_named_color(string::String)::Bool

Check if a string represents a named color.

Term.color.is_rgb_color — Method

is_rgb_color(string::String)::Bool

Check if a string represents a RGB color.