Colors

Term.ANSICode

Term.Colors.AbstractColor

Term.Colors.get_color

Term.Colors.hex2rgb

Term.Colors.is_background

Term.Colors.is_color

Term.Colors.is_hex_color

Term.Colors.is_named_color

Term.Colors.is_rgb_color

Term.Colors.nospaces

Term.ANSICode — Method

ANSICode(color; bg::Bool=false)

Create ANSI tags for colors.

Term.Colors.AbstractColor — Type

AbstractColor

Abstract color type.

Term.Colors.get_color — Method

get_color(string::String; bg=false)::AbstractColor

Extract a color type from a string with color information.

Term.Colors.hex2rgb — Method

hex2rgb(hex::String)

Converts a string hex color code to a RGB color

Term.Colors.is_background — Method

is_background(string::String)::Bool

Check if a string represents background color information, of any type.

Term.Colors.is_color — Method

is_color(string::String)::Bool

Check if a string represents color information, of any type.

Term.Colors.is_hex_color — Method

is_hex_color(string::String)::Bool

Check if a string represents a hex color.

Term.Colors.is_named_color — Method

is_named_color(string::String)::Bool

Check if a string represents a named color.

Term.Colors.is_rgb_color — Method

is_rgb_color(string::String)::Bool

Check if a string represents a RGB color.

Term.Colors.nospaces — Method

nospaces(text::AbstractString)

Remove all spaces from a string.