Consoles

Term.Consoles.beginning_previous_line

Term.Consoles.change_scroll_region

Term.Consoles.clear

Term.Consoles.console_height

Term.Consoles.console_width

Term.Consoles.cursor_position

Term.Consoles.erase_line

Term.Consoles.hide_cursor

Term.Consoles.line

Term.Consoles.move_to_line

Term.Consoles.next_line

Term.Consoles.restorecursor

Term.Consoles.savecursor

Term.Consoles.show_cursor

Term.Consoles.beginning_previous_line — Function

Move cursor to the beginning of the previous line

Term.Consoles.change_scroll_region — Function

Change the position of the scrolling region in the terminal.

See: http://www.sweger.com/ansiplus/EscSeqScroll.html

Term.Consoles.clear — Function

clear(io::IO = stdout)

Clear terminal from anything printed in the REPL.

Term.Consoles.console_height — Function

console_height()

Get the current console height.

Term.Consoles.console_width — Function

console_width()

Get the current console width.

Term.Consoles.cursor_position — Function

Get cursor position

Term.Consoles.erase_line — Function

Erase last line in console.

Term.Consoles.hide_cursor — Function

Hide cursor

Term.Consoles.line — Function

write a new line.

Term.Consoles.move_to_line — Function

Move cursor to a specific line

Term.Consoles.next_line — Function

Move cursor down one or more lines

Term.Consoles.restorecursor — Function

Restore a previously saved cursor position

Term.Consoles.savecursor — Function

Save the current cursor position

Term.Consoles.show_cursor — Function

Show cursor