Live

Term.LiveWidgets.AbstractMenu

Term.LiveWidgets.AbstractWidget

Term.LiveWidgets.AbstractWidgetContainer

Term.LiveWidgets.App

Term.LiveWidgets.App

Term.LiveWidgets.AppInternals

Term.LiveWidgets.Button

Term.LiveWidgets.ButtonsMenu

Term.LiveWidgets.Gallery

Term.LiveWidgets.InputBox

Term.LiveWidgets.MultiSelectMenu

Term.LiveWidgets.Pager

Term.LiveWidgets.PlaceHolderWidget

Term.LiveWidgets.SimpleMenu

Term.LiveWidgets.TextWidget

Term.LiveWidgets.ToggleButton

Term.LiveWidgets.WidgetInternals

Term.LiveWidgets.activate_next_widget

Term.LiveWidgets.activate_prev_widget

Term.LiveWidgets.active_title

Term.LiveWidgets.active_title

Term.LiveWidgets.add_debugging_info!

Term.LiveWidgets.addchar

Term.LiveWidgets.addspace

Term.LiveWidgets.del

Term.LiveWidgets.enforce_app_size

Term.LiveWidgets.enforce_app_size

Term.LiveWidgets.erase!

Term.LiveWidgets.execute_transition_rule

Term.LiveWidgets.frame

Term.LiveWidgets.frame

Term.LiveWidgets.frame

Term.LiveWidgets.get_active

Term.LiveWidgets.get_active

Term.LiveWidgets.home

Term.LiveWidgets.inactive_title

Term.LiveWidgets.inactive_title

Term.LiveWidgets.infer_transition_rules

Term.LiveWidgets.isactive

Term.LiveWidgets.keyboard_input

Term.LiveWidgets.keyboard_input

Term.LiveWidgets.make_button_panel

Term.LiveWidgets.make_page_content

Term.LiveWidgets.make_placeholders

Term.LiveWidgets.make_scrollbar

Term.LiveWidgets.menu_activate_next

Term.LiveWidgets.menu_activate_prev

Term.LiveWidgets.menu_return_value

Term.LiveWidgets.menu_return_value

Term.LiveWidgets.multi_select_toggle

Term.LiveWidgets.newline

Term.LiveWidgets.next_line

Term.LiveWidgets.next_page

Term.LiveWidgets.on_activated

Term.LiveWidgets.on_deactivated

Term.LiveWidgets.on_layout_change

Term.LiveWidgets.play

Term.LiveWidgets.press_button

Term.LiveWidgets.prev_line

Term.LiveWidgets.prev_page

Term.LiveWidgets.print_node

Term.LiveWidgets.quit

Term.LiveWidgets.refresh!

Term.LiveWidgets.replace_line

Term.LiveWidgets.replace_line

Term.LiveWidgets.reshape_pager_content

Term.LiveWidgets.set_active

Term.LiveWidgets.set_as_parent

Term.LiveWidgets.shouldupdate

Term.LiveWidgets.stop!

Term.LiveWidgets.toend

Term.LiveWidgets.toggle_help

Term.LiveWidgets.AbstractMenu — Type

AbstractMenu

Objects of type AbstractMenu present the user with a few options and let them select one by moving around with the up/down keys and using Enter to select an option.

Term.LiveWidgets.AbstractWidget — Type

AbstractWidget

Abstract widgets must have three obligatory fields: measure::Measure controls:: Dict{Union{KeyInput, Char}, Function} parent::Union{Nothing, AbstractWidget}

and one optional one on_draw::Union{Nothing, Function} = nothing

Term.LiveWidgets.AbstractWidgetContainer — Type

AbstractWidgetContainer

AbstractWidgetContainer contain multiple other widgets and can coordinate their activity. AbstractWidgetContainer must have two obligatory fields on top of AbstractWidget's fields: widgets::Vector{AbstractWidget} active::Int

Term.LiveWidgets.App — Type

An App is a collection of widgets.

Tip

Transition rules bind keys to "movement" in the app to change focus to a different widget

Term.LiveWidgets.App — Method

App(
    widget::AbstractWidget;
    controls::AbstractDict = app_controls,
    width=1.0,
    height=min(40, console_height()),
    kwargs...
)

Convenience constructor for an App with a single widget.

Term.LiveWidgets.AppInternals — Type

struct AppInternals iob::IOBuffer ioc::IOContext term::AbstractTerminal prevcontent::Union{Nothing, AbstractRenderable, String} prevcontentlines::Vector{String} rawmodeenabled::Bool lastupdate::Union{Nothing, Int} refreshΔt::Int help_shown::Bool end

AppInternals handles "under the hood" work for live widgets. It takes care of keeping track of information such as the content displayed at the last refresh of the widget to inform the printing of the widget's content at the next refresh.

AppInternals also holds linked_widgets which can be used to link to other widgets to access their internal variables.

Term.LiveWidgets.Button — Type

Button

A button widget. It's display changes when pressed. A callback can be set to be called when the button is pressed.

Term.LiveWidgets.ButtonsMenu — Type

Simple menu in which each option is a Panel object. Styling reflects which option is currently selected

Term.LiveWidgets.Gallery — Type

A Gallery containes multiple widgets, but only shows one at the time.

Term.LiveWidgets.InputBox — Type

InputBox collects and displays user input as text.

Term.LiveWidgets.MultiSelectMenu — Type

Menu variant for selecting multiple options. Color indicates current active option, ticks selected options

Term.LiveWidgets.Pager — Type

A Pager is a widget for visualizing long texts a few lines at the time. It shows a few lines of a longer text and allows users to move up and down the text using keys such as arrow up and arrow down.

Term.LiveWidgets.PlaceHolderWidget — Type

Widget with no content to be used as a placeholder for choosing app layout.

Term.LiveWidgets.SimpleMenu — Type

Simple text based menu. The currently selected option is highlighted with a different style.

Term.LiveWidgets.TextWidget — Type

TextWidget just shows a piece of text.

Term.LiveWidgets.ToggleButton — Type

A button. Pressing it toggles its status between activated and not.

Term.LiveWidgets.WidgetInternals — Type

WidgetInternals

This struct is used to store the internal state of a widget as well as callbacks assigned to it.

Term.LiveWidgets.activate_next_widget — Method

Set the active widget in container to the next widget.

Term.LiveWidgets.activate_prev_widget — Method

Set the active widget in container to the previous widget.

Term.LiveWidgets.active_title — Method

active_title(mn::ButtonsMenu, i::Int, width::Int, height::Int; panel_kwargs...)

Return the title of the i-th item of the menu with the active style.

Term.LiveWidgets.active_title — Method

active_title(mn::SimpleMenu, i::Int, width::Int)

Return the title of the i-th item of the menu with the active style.

Term.LiveWidgets.add_debugging_info! — Method

add_debugging_info!(content::AbstractRenderable, app::App)::AbstractRenderable

Add debugging information to the top of the app.

Term.LiveWidgets.addchar — Method

add character to input

Term.LiveWidgets.addspace — Method

insert space

Term.LiveWidgets.del — Method

delete last character

Term.LiveWidgets.enforce_app_size — Method

enforce_app_size(compositor::Compositor, widgets::AbstractDict)

Called when an App is first created to set the size of all widgets.

Term.LiveWidgets.enforce_app_size — Method

enforce_app_size(app::App, measure::Measure)

Called when a console is resized to adjust the apps layout.

Term.LiveWidgets.erase! — Method

erase!(app::App)

Erase a app from the terminal.

Term.LiveWidgets.execute_transition_rule — Method

Execute a transition rule to switch focus to another widget.

Term.LiveWidgets.frame — Method

frame(pager::Pager)::AbstractRenderable

Create a Panel with, as content, the currently visualized lines in the Pager.

Term.LiveWidgets.frame — Method

Get the current content of a widget

Term.LiveWidgets.frame — Method

frame(app::App)

Render the app and its content.

Term.LiveWidgets.get_active — Method

Return the active widget in container.

Term.LiveWidgets.get_active — Method

get_active(w::AbstractWidget)

Nothing: no children.

Term.LiveWidgets.home — Method

move to first line

Term.LiveWidgets.inactive_title — Method

inactive_title(mn::ButtonsMenu, i::Int, width::Int, height::Int; panel_kwargs...)

Return the title of the i-th item of the menu with the inactive style.

Term.LiveWidgets.inactive_title — Method

inactive_title(mn::SimpleMenu, i::Int, width::Int)

Return the title of the i-th item of the menu with the inactive style.

Term.LiveWidgets.infer_transition_rules — Method

infer_transition_rules(layout::Expr)::Dict

If no transition rules are passed, infer them from the layout's spatial relationships.

Term.LiveWidgets.isactive — Method

isactive(w::AbstractWidget)

Returns true if the widget is active, i.e. if it is the active widget

Term.LiveWidgets.keyboard_input — Method

keyboard_input(widget::AbstractWidgetContainer)

Get keyboard input from the terminal and execute the corresponding control function for active widgets.

Term.LiveWidgets.keyboard_input — Method

keyboard_input(widget::AbstractWidget)

Get keyboard input from the terminal and execute the corresponding control function if it exists. Returns a list of return values from the control functions.

Term.LiveWidgets.make_button_panel — Method

make_button_panel(message, color, text_color, pressed, active, w, h; kwargs...)

Create a panel to display a button.

Term.LiveWidgets.make_page_content — Method

make_page_content(pager::Pager, i::Int, Δi::Int)::String

Returns the content of the page as a string

Term.LiveWidgets.make_placeholders — Method

If no widget was passed, create placeholder widgets.

Term.LiveWidgets.make_scrollbar — Method

make_scrollbar(pager::Pager, i::Int, Δi::Int)::String

Generate a scrollbar display to show progress in the pager's content.

Term.LiveWidgets.menu_activate_next — Method

set next item to active

Term.LiveWidgets.menu_activate_prev — Method

set previous item to active

Term.LiveWidgets.menu_return_value — Method

quit menu and return value

Term.LiveWidgets.menu_return_value — Method

quit and return selection.

Term.LiveWidgets.multi_select_toggle — Method

Toggle selection status of current active option

Term.LiveWidgets.newline — Method

new line

Term.LiveWidgets.next_line — Method

move to the next line

Term.LiveWidgets.next_page — Method

move to the next page

Term.LiveWidgets.on_activated — Method

Default callback for a widget being activated

Term.LiveWidgets.on_deactivated — Method

Default callback for a widget being deactivated

Term.LiveWidgets.on_layout_change — Method

on_layout_change(app::App)

Called when the console is resized to adjust the apps layout.

Term.LiveWidgets.play — Method

play(app::App; transient::Bool=true)

Keep refreshing a renderable, until the user interrupts it.

Term.LiveWidgets.press_button — Method

set the buttoon's state to :presed.

Term.LiveWidgets.prev_line — Method

move to the previous line

Term.LiveWidgets.prev_page — Method

move to the previous page

Term.LiveWidgets.print_node — Method

print_node(io, x)

Print function to print a node (widget) in an application's hierarchy tree. It prints the node's stated dimensions vs its content's (calling frame). Used for debugging

Term.LiveWidgets.quit — Function

Quit the current app, potentially returning some value.

Term.LiveWidgets.refresh! — Method

refresh!(live::AbstractWidget)::Tuple{Bool, Any}

Update the terminal display of a app.

this is done by calling frame on the app to get the new content. Then, line by line, the new content is compared to the previous one and when a discrepancy occurs the lines gets re-written. This is all done printing to a buffer first and then to stdout to avoid jitter.

Term.LiveWidgets.replace_line — Method

replace_line(internals::AppInternals, newline)

Erase a line, write new content and move cursor.

Term.LiveWidgets.replace_line — Method

replace_line(internals::AppInternals)

Erase a line and move cursor

Term.LiveWidgets.reshape_pager_content — Method

reshape_pager_content(content::AbstractString, line_numbers::Bool)::Vector{string}

Turns a text into a vector of lines with the right size (and optionally line numbers)

Term.LiveWidgets.set_active — Method

Set the active widget in container to active.

Term.LiveWidgets.set_as_parent — Method

Set the parent of all widgets in container to container.

Term.LiveWidgets.shouldupdate — Method

shouldupdate(app::App)::Bool

Check if a widget's display should be updated based on: 1. enough time elapsed since last update 2. the widget has not beed displayed het

Term.LiveWidgets.stop! — Method

stop!(app::App)

Restore normal terminal behavior.

Term.LiveWidgets.toend — Method

move to the last line

Term.LiveWidgets.toggle_help — Method

display/hide help tooltip