Logging

Term.logging.TermLogger

Logging.handle_message

Term.logging.install_term_logger

Term.logging.print_closing_line

Term.logging.TermLogger — Type

TermLogger

Custom logger type.

Logging.handle_message — Method

Logging.handle_message(logger::TermLogger,

Handle printing of log messages, with style!.

In addition to the log message and info such as file/line and time of log, it prints kwargs styled by their type.

Term.logging.install_term_logger — Function

install_term_logger(theme::Theme=theme)

Install TermLogger as the global logging system.

theme::Theme can be passed to specify the theme to use for styling objects.

Term.logging.print_closing_line — Function

print_closing_line(color::String, width::Int)

Print the final line of a log message with style and date info