Logs

Term.Logs.TermLogger

Logging.handle_message

Term.Logs.handle_progress

Term.Logs.install_term_logger

Term.Logs.log_value_display

Term.Logs.print_closing_line

Term.Logs.TermLogger — Type

TermLogger

Custom logger type.

Logging.handle_message — Method

handle_message(logger::TermLogger, lvl, msg, _mod, group, id, file, line; kwargs...)

Handle printing of log messages, with style!.

In addition to the log message and info such as file/line and time of log, it prints kwargs styled by their type.

Term.Logs.handle_progress — Method

handle_progress(logger::TermLogger, prog)

Handle progress information passed by ProgressLogging.

It creates/adds/removes ProgressJobs to the logger's ProgressBar to create progress visualizations.

Term.Logs.install_term_logger — Function

install_term_logger(theme::Theme=theme)

Install TermLogger as the global logging system.

theme::Theme can be passed to specify the theme to use for styling objects.

Term.Logs.log_value_display — Function

Create string display for a log message value.

Term.Logs.print_closing_line — Function

print_closing_line(color::String, width::Int)

Print the final line of a log message with style and date info