Measure

Term.measure.Measure

Term.measure.Measure

Base.:+

Term.measure.height

Term.measure.width

Term.measure.Measure — Type

Measure

Stores the size of a piece of renderable material

Term.measure.Measure — Method

Measure(str::String)

Constructs a measure object from a string

Base.:+ — Method

The sum of measures returns a measure with the highest value along each dimension

Term.measure.height — Function

height

Measure the height of renderable objects (text, AbsstractRenderable)

Term.measure.width — Function

width

Measure the width of renderable objects (text, AbsstractRenderable)