Prompts

Term.Prompts

Term.Prompts.AbstractDefaultPrompt

Term.Prompts.AbstractOptionsPrompt

Term.Prompts.AbstractPrompt

Term.Prompts.AnswerValidationError

Term.Prompts.DefaultPrompt

Term.Prompts.OptionsPrompt

Term.Prompts.Prompt

Term.Prompts.TypePrompt

Base.print

Base.print

Base.print

Term.Prompts.ask

Term.Prompts.ask

Term.Prompts.ask

Term.Prompts.validate_answer

Term.Prompts.validate_answer

Term.Prompts.validate_answer

Term.Prompts — Module

module Prompts

Defines functionality relative to prompts in the terminal. Typically a prompt is composed of a piece of text that gets displayed prompting the user to provide an input and some machinery to parse/validate the user's inputs. For example, some prompts may only accept as replies objects of a given type (e.g. an Int). Additionally, some prompts will have "options" the user can choose between and the answer has to be one of these options.

Term.Prompts.AbstractDefaultPrompt — Type

Options prompt types with a default answer

Term.Prompts.AbstractOptionsPrompt — Type

Prompt types where user can only choose among options

Term.Prompts.AbstractPrompt — Type

Prompt types

Term.Prompts.AnswerValidationError — Type

AnswerValidationError <: Exception

Exception to handle cases in which the user's answer to a prompt failed to pass validation.

Term.Prompts.DefaultPrompt — Type

Term.Prompts.OptionsPrompt — Type

struct OptionsPrompt <: AbstractOptionsPrompt
    options::Vector{String}
    prompt::String
    style::String 
    answers_style::String
end

Just a simple prompt, giving some pre-defined options.

Term.Prompts.Prompt — Type

struct Prompt{T} <: AbstractPrompt
    prompt::String
    style::String = TERM_THEME[].prompt_text
end

Generic prompt, accepts any answer

Term.Prompts.TypePrompt — Type

struct TypePrompt{T}
    answer_type::Union{Union, DataType} = T
    prompt::String
end

Asks for input given prompt and checks/converts the answer to type T

Base.print — Method

Base.print(io::IO, prompt::AbstractDefaultPrompt)

Print a prompt with style applied to the default option.

Base.print — Method

Base.print(io::IO, prompt::AbstractOptionsPrompt)

Options prompts additionally print the available options.

Base.print — Method

Base.print(io::IO, prompt::AbstractPrompt)

Default prompt printing, just prints the message prompt with a bit of style.

Term.Prompts.ask — Function

ask

Ask does three things:

displays a prompt

accepts user input and validates it

if the answer was accepted, returns the desired value.

Term.Prompts.ask — Method

ask(io::IO, prompt::AbstractOptionsPrompt)

In asking an AbstractOptionsPrompt, keep asking for input until an accepted answer is provided.

Term.Prompts.ask — Method

ask(io::IO, prompt::AbstractPrompt)

Default ask method for generic prompt objects.

Term.Prompts.validate_answer — Function

validate_answer

Validate user's answer for a prompt type. The validation mechanism depends on the type of prompt. Validate answer will return the answer if it passed validation or raise and error otherwise.

Term.Prompts.validate_answer — Method

validate_answer(answer, prompt::AbstractOptionsPrompt)

For an AbstractOptionsPrompt an answer is accepted if its one of the options. Additionally, for an AbstractDefaultPrompt, if no answer is given that's also accepted and the default option is returned.

Term.Prompts.validate_answer — Method

validate_answer(answer, prompt::TypePrompt)

For a TypePrompt an anwer is valid if it is of the correct type or if a string containg the answer can be parsed as the correct type. For example, answer="1.0" can be accepted for a TypePrompt asking for a Number. If validation fails, an error is raised.