REPR

Term.Repr.get_docstring

Term.Repr.get_methods_with_docstrings

Term.Repr.install_term_repr

Term.Repr.repr_panel

Term.Repr.style_function_methods

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.termshow

Term.Repr.typename

Term.Repr.with_repr

Term.Repr.@with_repr

Term.Repr.get_docstring — Method

get_docstring(obj)

Extract and style an object's docstring.

Term.Repr.get_methods_with_docstrings — Method

get_methods_with_docstrings(obj::Union{Union, DataType, Function})

Get the docstring for each method for an object (function/datatype).

Term.Repr.install_term_repr — Method

install_term_repr()

Make term_show be called every times something is shown in the REPL

Term.Repr.repr_panel — Method

Create a Panel showing repr content using the current theme's style info.

Term.Repr.style_function_methods — Method

Create a styled list of methods for a function. Accepts string(methods(function)) as argument.

Term.Repr.termshow — Function

termshow

Styled string representation of any object.

termshow prints to stdout (or any other IO) a styled representation of the object. Dedicated methods create displays for specify types such as Dict or Vector.

Term.Repr.termshow — Method

termshow(io::IO, arr::AbstractArray; kwargs...)

Show the content of a multidimensional array as a series of 2D slices.

Term.Repr.termshow — Method

termshow(io::IO, d::Dict; kwargs...)

Show a dictionary's keys and values and their data types.

Term.Repr.termshow — Method

termshow(io::IO, mtx::AbstractMatrix; kwargs...)

Show a matrix content as a 2D table visualization.

Term.Repr.termshow — Method

termshow(io::IO, obj)

Generic method for any object not caught by dedicated methods. Creates a Panel with the object's fields and contents.

Term.Repr.termshow — Method

termshow(io::IO, obj::DataType; kwargs...)

Show a type's arguments, constructors and docstring.

Term.Repr.termshow — Method

termshow(io::IO, e::Expr; kwargs...)

Show an expression's head and arguments.

Term.Repr.termshow — Method

termshow(io::IO, fun::Function; kwargs...)

Show a function's methods and docstring.

Term.Repr.termshow — Method

termshow(io::IO, vec::Union{Tuple,AbstractVector}; kwargs...)

Show a vector's content as a 1D table visualization.

Term.Repr.typename — Method

typename(typedef::Expr)

Get the name of a type as an expression

Term.Repr.with_repr — Method

with_repr(typedef::Expr)

Function for the macro @with_repr which creates a Base.show method for a type.

The show method shows the field names/types for the type and the values of the fields.

Example

@with_repr struct TestStruct2
    x::Int
    name::String
    y
end

Term.Repr.@with_repr — Macro

with_repr(typedef::Expr)

Function for the macro @with_repr which creates a Base.show method for a type.

The show method shows the field names/types for the type and the values of the fields.

Example

@with_repr struct TestStruct2
x::Int
name::String
y
end