Segment

Term.segment.Segment

Term.segment.Segment

Term.segment.Segment

Base.:*

Base.show

Term.fillin

Term.segment.Segment — Type

Segment

stores one piece of text with all the styles applied to it.

Term.segment.Segment — Method

Segment(text::Union{Segment, AbstractString}, markup::AbstractString)

Construct a Segment out of a plain string and a markup string with style info

Term.segment.Segment — Method

Segment(text::AbstractString)

Construct a Segment out of a string with markup.

Base.:* — Method

concatenate strings and segments

Base.show — Method

print styled in stdout, info otherwise

Term.fillin — Method

Term.fillin(segments::Vector{Segment})::Vector{Segment}

Ensure that for each segment the text has the same width