Segments

Term.Segments.Segment

Term.Segments.Segment

Term.Segments.Segment

Base.:*

Term.fillin

Term.Segments.Segment — Type

Segment

stores one piece of text with all the styles applied to it.

Term.Segments.Segment — Method

Segment(text::Union{Segment, AbstractString}, markup::AbstractString)

Construct a Segment out of a plain string and a markup string with style info

Term.Segments.Segment — Method

Segment(text::AbstractString)

Construct a Segment out of a string with markup.

Base.:* — Method

concatenate strings and segments

Term.fillin — Method

Term.fillin(segments::Vector{Segment})::Vector{Segment}

Ensure that for each segment the text has the same width