Style

Term.style.MarkupStyle

Term.style.MarkupStyle

Term.style.apply_style

Term.style.apply_style

Term.style.apply_style

Term.style.extract_style

Term.style.get_style_codes

Term.style.MarkupStyle — Type

MarkupStyle

Holds information about the style specification set out by a MarkupTag.

Term.style.MarkupStyle — Method

MarkupStyle(tag::MarkupTag)

Builds a MarkupStyle definition from a MarkupTag.

Term.style.apply_style — Method

apply_style(text::AbstractString, tag::MarkupTag)::AbstractString

Appliy the style of a markup tag and it's nested tags

Term.style.apply_style — Method

apply_style(text::AbstractString, style::MarkupStyle)::AbstractString

Apply a style to a text.

Term.style.apply_style — Method

apply_style(text::AbstractString)

Extract and apply all markup style in a string.

Term.style.extract_style — Method

extract_style(text::AbstractString)

Extract style information from a text.

Term.style.get_style_codes — Method

get_style_codes(style::MarkupStyle)

Get ANSICodes corresponding to a MarkupStyle.