Layout

Term.Tprint.tprint

Term.Tprint.tprint

Term.Tprint.tprint

Term.Tprint.tprint

Term.Tprint.tprint

Term.Tprint.tprint

Term.Tprint.tprint

Term.Tprint.tprint

Term.Tprint.tprint — Function

tprint

Similar to standard lib's print function but with added styling functionality

Term.Tprint.tprint — Method

tprint(x)

When no dedicated method is present, print the string representation

Term.Tprint.tprint — Method

tprint(x::DataType)

Print highlighted as a DataType

Term.Tprint.tprint — Method

tprint(x::Function)

Print highlighted as a Function

Term.Tprint.tprint — Method

tprint(x::AbstractString)

Apply style to a string and print it to a new line

Term.Tprint.tprint — Method

tprint(x::Number)

Print highlighted as a Number

Term.Tprint.tprint — Method

tprint(x::Symbol)

Print highlighted as a Symbol

Term.Tprint.tprint — Method

tprint(x::AbstractRenderable)

Print an AbstractRenderable.

Equivalent to println(x)