Tree

Term.tree.Leaf

Term.tree.Tree

Term.tree.Tree

Term.tree.Tree

Term.tree.TreeGuides

Term.tree.TreeGuides

Base.show

Term.tree.addnode!

Term.tree.asleaf

Term.tree.make_hierarchy_dict

Term.tree.render

Term.tree.style_T

Term.tree.Leaf — Type

Leaf

End items in a Tree. No sub-trees.

Term.tree.Tree — Type

Tree

A tree is composed of nodes (other trees) and leaves (end items). It renders as a hierarchical structure with lines (guides) connecting the various elements

Term.tree.Tree — Method

Tree(T::DataType)::Tree

Construct a Tree visualization of T's types hierarchy

Term.tree.Tree — Method

Tree(data::Union{Dict, Pair}; level=0, title::String="tree", kwargs...)

Construct a Tree out of a Dict. Recursively handle nested Dicts.

Term.tree.TreeGuides — Type

TreeGuides

Store strings to make up a Tree's guides (the line elements showing connections).

Term.tree.TreeGuides — Method

TreeGuides(guides_type::Symbol, style::String)

Get tree guides with style information applied

Base.show — Method

Show/render a Tree

Term.tree.addnode! — Method

Add a new node to an existing tree's nodes or levaes.

Term.tree.asleaf — Function

asleaf

Style an object to render it as a a string

Term.tree.make_hierarchy_dict — Method

make_hierarchy_dict(x::Vector{DataType}, T::DataType, Tsubs::Dict)::Dict

Recursively create a dictionary with the types hierarchy for T. Tsubs carries information about T's subtypes. The Dict is made backwards. From the deepest levels up.

Term.tree.render — Method

render(tree::Tree)::Vector{Segment}

Render a Tree into segments. Recursively handle nested trees.

Term.tree.style_T — Method

Apply style for the type whose hierarchy Tree we are making