Trees

Term.Trees.Leaf

Term.Trees.Tree

Term.Trees.Tree

Term.Trees.Tree

Term.Trees.TreeGuides

Term.Trees.TreeGuides

Base.show

Term.Trees.addnode!

Term.Trees.asleaf

Term.Trees.make_hierarchy_dict

Term.Trees.render

Term.Trees.style_T

Term.Trees.Leaf — Type

Leaf

End items in a Tree. No sub-trees.

Term.Trees.Tree — Type

Tree

A tree is composed of nodes (other trees) and leaves (end items). It renders as a hierarchical structure with lines (guides) connecting the various elements

Term.Trees.Tree — Method

Tree(T::DataType)::Tree

Construct a Tree visualization of T's types hierarchy

Term.Trees.Tree — Method

Tree(data::Union{AbstractDict, Pair}; level=0, title::String="tree", kwargs...)

Construct a Tree out of a Dict or OrderedDict. Recursively handle nested Dicts.

Term.Trees.TreeGuides — Type

TreeGuides

Store strings to make up a Tree's guides (the line elements showing connections).

Term.Trees.TreeGuides — Method

TreeGuides(guides_type::Symbol, style::String)

Get tree guides with style information applied

Base.show — Method

Show/render a Tree

Term.Trees.addnode! — Method

Add a new node to an existing tree's nodes or levaes.

Term.Trees.asleaf — Function

asleaf

Style an object to render it as a a string

Term.Trees.make_hierarchy_dict — Method

make_hierarchy_dict(x::Vector{DataType}, T::DataType, Tsubs::AbstractDict)::AbstractDict

Recursively create a dictionary with the types hierarchy for T. Tsubs carries information about T's subtypes. The AbstractDict is made backwards. From the deepest levels up.

Term.Trees.render — Method

render(tree::Tree)::Vector{Segment}

Render a Tree into segments. Recursively handle nested trees.

Term.Trees.style_T — Method

Apply style for the type whose hierarchy Tree we are making