Term.box.Box — Type

Box

Defines characters to render boxes.

Row names:

┌─┬┐ top │ ││ head ├─┼┤ headrow │ ││ mid ├─┼┤ row ├─┼┤ footrow │ ││ foot └─┴┘ bottom

each row is an instance of BoxLine

Term.box.Box — Method

Box(string)

Construct a Box objet out of a box string.

Term.box.BoxLine — Type

BoxLine

Stores the characters for a line of a Box object.

Term.box.chars — Method

Split a string into a vector of Chars.

Term.box.cleantext — Method

cleantext(str::AbstractString)

Remove all style information from a string.

Term.box.do_by_line — Method

do_by_line(fn::Function, text::AbstractString)

Apply fn to each line in the text.

The function fn should accept a single ::AbstractString argument.

Term.box.escape_brackets — Method

remove_ansi(str::AbstractString)::AbstractString

Replace each squared bracket with a double copy of itself

Term.box.fit — Method

fit(box::Box, widths::Vector{Int})::String

Creates a box.

The box has one of each level type with columns widths specified by a vector of widhts.

Term.box.get_last_valid_str_idx — Method

get_last_valid_str_idx(str::AbstractString, idx::Int)

Get valid index to cut a string at.

When indexing a string, the number of indices is given by the the sum of the ncodeunits of each Char, but some indices will not be valid. This function ensures that given a (potentially) not valid index, the last valid one is elected.

Term.box.get_next_valid_str_idx — Method

getnextvalidstridx(str::AbstractString, idx::Int)

Get valid index to cut a string at.

When indexing a string, the number of indices is given by the the sum of the ncodeunits of each Char, but some indices will not be valid. This function ensures that given a (potentially) not valid index, the next valid one is elected.

Term.box.get_row — Method

get_row(box, [1, 2, 3], :row)

Gets characters for a row of a Box object.

The level Symbol can be used to specify the box level (:top, :footer...)

Term.box.get_title_row — Method

gettitlerow(row::Symbol, box::Box, title::Union{Nothing, AbstractString}; <keyword arguments>)

Create a box row with a title string.

Can create both titles in the top and bottom row to produce subtitles.

#Arguments:

width::Int: width of line

style::Union{Nothing: String}: style of line

title_style::Union{Nothing: AbstractString}: style of title string

justify::Symbol=:left: position of title string

See also get_row.

Term.box.get_valid_chars! — Method

get_valid_chars!(valid_chars::Vector{Int}, tag, δ::Int)

Recursively extract valid characters (i.e. not in markup tags) from a string.

Term.box.join_lines — Method

join_lines(lines)

Merge a vector of strings in a single string.

Term.box.loop_last — Method

loop_last(v::Vector)

Returns an iterable yielding tuples (is_last, value).

Term.box.nospaces — Method

nospaces(text::AbstractString)

Remove all spaces from a string.

Term.box.read_file_lines — Method

read_file_lines(path::String, start::Int, stop::Int)

Read a file and select only lines in range start -> stop.

Returns a vector of tuples with the line number and line content.

Term.box.reinsert_double_brackets — Method

reinsert_double_brackets(text::AbstractString)::AbstractString

Insert previously replaced double brackets

Term.box.remove_ansi — Method

remove_ansi(str::AbstractString)::AbstractString

Remove all ANSI tags from a string of text

Term.box.remove_brackets — Method

remove_brackets(text::AbstractString)

Remove all () brackets from a string.

Term.box.remove_markup — Method

remove_markup(input_text::AbstractString)::AbstractString

Remove all markup tags from a string of text.

Term.box.remove_markup_open — Method

remove_markup_open(text::AbstractString)

Remove all opening markup tags from a string of text

Term.box.replace_double_brackets — Method

replace_double_brackets(text::AbstractString)::AbstractString

Replace double brackets with %% and ±± to avoid them being picked up by markup extraction

Term.box.replace_text — Function

replace_text(text::AbstractString, start::Int, stop::Int, char::Char='_')

Replace a section of a text between start and stop with another string composed of repeats of a given character char.

Term.box.replace_text — Method

replace_text(text::AbstractString, start::Int, stop::Int, replace::AbstractString)

Replace a section of a text between start and stop with replace.

Term.box.reshape_text — Method

reshape_text(text::AbstractString, width::Int)

Reshape text to have a given width.

When text is longer than width, it gets cut into multiple lines. This is done carefully to preserve style information by: avoiding cutting inside style markup and copying markup tags over to new lines so that the style is correctly applied.

Term.box.split_lines — Method

split_lines(text::AbstractString)

Split a string into its composing lines.

Term.box.split_lines — Method

split_lines(renderable)

Split a renderable's text.

Term.box.square_to_round_brackets — Method

square_to_round_brackets(text::AbstractString)

Replace square brackets with round ones.

Term.box.textlen — Method

textlen(x::AbstractString)

Get length of text after all style information is removed.

Term.box.truncate — Method

truncate(text::AbstractString, width::Int)

Shorten a string of text to a target width

Term.box.unescape_brackets — Method

unescape_brackets(text::AbstractString)::AbstractString

Replece every double squared parenthesis with a single copy of itself

Term.box.unspace_commas — Method

unspace_commas(text::AbstractString)

Remove spaces after commas.