Term.layout.Padding — Type

Padding

Stores empty string to pad a string to the desired with.

Term.layout.Padding — Method

Padding(text::AbstractString, target_width::Int, method::Symbol)::Padding

Create Padding for a string given a justification method ∈ (:left, :center, :right).

Term.layout.Spacer — Type

    Spacer

A box of empty text with given width and height.

Term.layout.hLine — Type

hLine

A 1-line renderable made of repeated character from a Box.

Term.layout.hLine — Method

hLine(text::AbstractString; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Construct an hLine as wide as the stdout with centered text.

Term.layout.hLine — Method

hLine(width::Number, text::String; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Creates an hLine object with texte centered horizontally.

Term.layout.hLine — Method

hLine(width::Number, style::Union{String, Nothing}; box::Symbol=:ROUNDED)

Create a styled hLine of given width.

Term.layout.hLine — Method

hLine(; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Construct an hLine as wide as the stdout

Term.layout.vLine — Type

vLine

A multi-line renderable with each line made of a | to create a vertical line

Term.layout.vLine — Method

vLine(height::Number, style::Union{String, Nothing}; box::Symbol=:ROUNDED)

Create a vLine given a height and, optionally, style information.

Term.layout.vLine — Method

vLine(; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Create a vLine as tall as the stdout console

Term.layout.chars — Method

Split a string into a vector of Chars.

Term.layout.cleantext — Method

cleantext(str::AbstractString)

Remove all style information from a string.

Term.layout.do_by_line — Method

do_by_line(fn::Function, text::AbstractString)

Apply fn to each line in the text.

The function fn should accept a single ::AbstractString argument.

Term.layout.escape_brackets — Method

remove_ansi(str::AbstractString)::AbstractString

Replace each squared bracket with a double copy of itself

Term.layout.get_last_valid_str_idx — Method

get_last_valid_str_idx(str::AbstractString, idx::Int)

Get valid index to cut a string at.

When indexing a string, the number of indices is given by the the sum of the ncodeunits of each Char, but some indices will not be valid. This function ensures that given a (potentially) not valid index, the last valid one is elected.

Term.layout.get_next_valid_str_idx — Method

getnextvalidstridx(str::AbstractString, idx::Int)

Get valid index to cut a string at.

When indexing a string, the number of indices is given by the the sum of the ncodeunits of each Char, but some indices will not be valid. This function ensures that given a (potentially) not valid index, the next valid one is elected.

Term.layout.get_valid_chars! — Method

get_valid_chars!(valid_chars::Vector{Int}, tag, δ::Int)

Recursively extract valid characters (i.e. not in markup tags) from a string.

Term.layout.hstack — Method

hstack(renderables...)

Horizonatlly stack a variable number of renderables

Term.layout.hstack — Method

hstack(r1::RenderablesUnion, r2::RenderablesUnion)

Horizontally stack two renderables to give a new renderable.

Term.layout.join_lines — Method

join_lines(lines)

Merge a vector of strings in a single string.

Term.layout.nospaces — Method

nospaces(text::AbstractString)

Remove all spaces from a string.

Term.layout.read_file_lines — Method

read_file_lines(path::String, start::Int, stop::Int)

Read a file and select only lines in range start -> stop.

Returns a vector of tuples with the line number and line content.

Term.layout.reinsert_double_brackets — Method

reinsert_double_brackets(text::AbstractString)::AbstractString

Insert previously replaced double brackets

Term.layout.remove_ansi — Method

remove_ansi(str::AbstractString)::AbstractString

Remove all ANSI tags from a string of text

Term.layout.remove_brackets — Method

remove_brackets(text::AbstractString)

Remove all () brackets from a string.

Term.layout.remove_markup — Method

remove_markup(input_text::AbstractString)::AbstractString

Remove all markup tags from a string of text.

Term.layout.remove_markup_open — Method

remove_markup_open(text::AbstractString)

Remove all opening markup tags from a string of text

Term.layout.replace_double_brackets — Method

replace_double_brackets(text::AbstractString)::AbstractString

Replace double brackets with %% and ±± to avoid them being picked up by markup extraction

Term.layout.replace_text — Function

replace_text(text::AbstractString, start::Int, stop::Int, char::Char='_')

Replace a section of a text between start and stop with another string composed of repeats of a given character char.

Term.layout.replace_text — Method

replace_text(text::AbstractString, start::Int, stop::Int, replace::AbstractString)

Replace a section of a text between start and stop with replace.

Term.layout.reshape_text — Method

reshape_text(text::AbstractString, width::Int)

Reshape text to have a given width.

When text is longer than width, it gets cut into multiple lines. This is done carefully to preserve style information by: avoiding cutting inside style markup and copying markup tags over to new lines so that the style is correctly applied.

Term.layout.split_lines — Method

split_lines(text::AbstractString)

Split a string into its composing lines.

Term.layout.split_lines — Method

split_lines(renderable)

Split a renderable's text.

Term.layout.square_to_round_brackets — Method

square_to_round_brackets(text::AbstractString)

Replace square brackets with round ones.

Term.layout.textlen — Method

textlen(x::AbstractString)

Get length of text after all style information is removed.

Term.layout.truncate — Method

truncate(text::AbstractString, width::Int)

Shorten a string of text to a target width

Term.layout.unescape_brackets — Method

unescape_brackets(text::AbstractString)::AbstractString

Replece every double squared parenthesis with a single copy of itself

Term.layout.unspace_commas — Method

unspace_commas(text::AbstractString)

Remove spaces after commas.

Term.layout.vstack — Method

vstack(renderables...)

Vertically stack a variable number of renderables

Term.layout.vstack — Method

vstack(r1::RenderablesUnion, r2::RenderablesUnion)

Vertically stack two renderables to give a new renderable.