Color

Term.color.AbstractColor

Term.color._rgb

Term.color._rgb

Term.color.get_color

Term.color.hex2rgb

Term.color.is_background

Term.color.is_color

Term.color.is_hex_color

Term.color.is_named_color

Term.color.is_rgb_color

Term.color.AbstractColor — Type

AbstractColor

Abstract color type.

Term.color._rgb — Method

_rgb(numbertype, txt)

Tries to parse r,g,b out of a string based on number type.

Term.color._rgb — Method

_rgb(numbertype, txt)

Tries to parse r,g,b out of a string.

Term.color.get_color — Method

get_color(string::AbstractString; bg=false)::AbstractColor

Extract a color type from a string with color information.

Term.color.hex2rgb — Method

hex2rgb(hex::AbstractString)

Converts a string hex color code to a RGB color

Term.color.is_background — Method

is_background(string::AbstractString)::Bool

Check if a string represents background color information, of any type.

Term.color.is_color — Method

is_color(string::AbstractString)::Bool

Check if a string represents color information, of any type.

Term.color.is_hex_color — Method

is_hex_color(string::AbstractString)::Bool

Check if a string represents a hex color.

Term.color.is_named_color — Method

is_named_color(string::AbstractString)::Bool

Check if a string represents a named color.

Term.color.is_rgb_color — Method

is_rgb_color(string::AbstractString)::Bool

Check if a string represents a RGB color.