Console

Term.console.beginning_previous_line

Term.console.change_scroll_region

Term.console.clear

Term.console.console_height

Term.console.console_width

Term.console.cursor_position

Term.console.erase_line

Term.console.hide_cursor

Term.console.line

Term.console.move_to_line

Term.console.next_line

Term.console.restorecursor

Term.console.savecursor

Term.console.show_cursor

Term.console.beginning_previous_line — Function

Move cursor to the beginning of the previous line

Term.console.change_scroll_region — Function

Change the position of the scrolling region in the terminal.

See: http://www.sweger.com/ansiplus/EscSeqScroll.html

Term.console.clear — Function

Clear terminal.

Term.console.console_height — Function

console_height()

Get the current console height.

Term.console.console_width — Function

console_width()

Get the current console width.

Term.console.cursor_position — Function

Get cursor position

Term.console.erase_line — Function

Erase last line in console.

Term.console.hide_cursor — Function

Hide cursor

Term.console.line — Function

write a new line.

Term.console.move_to_line — Function

Move cursor to a specific line

Term.console.next_line — Function

Move cursor down one or more lines

Term.console.restorecursor — Function

Restore a previously saved cursor position

Term.console.savecursor — Function

Save the current cursor position

Term.console.show_cursor — Function

Show cursor