Console

Term.consoles.Console

Term.consoles.beginning_previous_line

Term.consoles.clear

Term.consoles.console_height

Term.consoles.console_width

Term.consoles.cursor_position

Term.consoles.down

Term.consoles.erase_line

Term.consoles.hide_cursor

Term.consoles.line

Term.consoles.show_cursor

Term.consoles.up

Term.consoles.Console — Type

Console

The Console object stores information about the dimensions of the output(::IO) where objects will be printed

Term.consoles.beginning_previous_line — Method

Move cursor to the beginning of the previous line

Term.consoles.clear — Method

Clear terminal.

Term.consoles.console_height — Method

console_height()

Get the current console height.

Term.consoles.console_width — Method

console_width()

Get the current console width.

Term.consoles.cursor_position — Method

Get cursor position

Term.consoles.down — Method

Move cursor down one line

Term.consoles.erase_line — Method

Erase last line in console.

Term.consoles.hide_cursor — Method

Hide cursor

Term.consoles.line — Method

Print a new line.

Term.consoles.show_cursor — Method

Show cursor

Term.consoles.up — Method

Move cursor up one line