Consoles

Term.console.Consoles

Term.console.console_height

Term.console.console_width

Term.console.tprint

Term.console.tprint

Term.console.tprint

Term.console.tprint

Term.console.tprint

Term.console.tprint

Term.console.tprint

Term.console.Consoles — Type

Consoles

The Consoles object stores information about the dimensions of the output(::IO) where objects will be printed

Term.console.console_height — Method

console_height()

Get the current console height.

Term.console.console_width — Method

console_width()

Get the current console width.

Term.console.tprint — Function

tprint

Similar to standard lib's print function but with added styling functionality

Term.console.tprint — Method

tprint(x::AbstractString)

Apply style to a string and print it to a new line

Term.console.tprint — Method

tprint(x)

When no dedicated method is present, print the string representation

Term.console.tprint — Method

tprint(x::Function)

Print highlighted as a Function

Term.console.tprint — Method

tprint(x::Number)

Print highlighted as a Number

Term.console.tprint — Method

tprint(x::Symbol)

Print highlighted as a Symbol

Term.console.tprint — Method

tprint(x::AbstractRenderable)

Print an AbstractRenderable.

Equivalent to println(x)