Errors

Term.errors._highlight

Term.errors._highlight_numbers

Term.errors._highlight_with_type

Term.errors.backtrace_subpanel

Term.errors.style_backtrace

Term.errors.style_error

Term.errors.style_stacktrace_simple

Term.errors._highlight — Function

_highlight(x)

Apply style to x based on its type.

Term.errors._highlight_numbers — Method

_highlight_numbers(x::AbstractString)

Add style to each number in a string

Term.errors._highlight_with_type — Method

_highlight_with_type(x)

Apply style to x and and mark its type.

Term.errors.backtrace_subpanel — Method

backtrace_subpanel(line::String, WIDTH::Int, title::String)

Create a subpanel for stacktrace, showing source code.

Term.errors.style_backtrace — Method

style_backtrace(io::IO, t::Vector)

Create a Panel with styled error backtrace information.

Term.errors.style_error — Method

style_error(io::IO, er)

Create a style error panel.

Creates a Panel with an error message and optional hints.

Term.errors.style_stacktrace_simple — Method

style_stacktrace_simple(stack::Vector)

Simply style a stacktrace. Just adding a miniumum of color.