Introspection

Term.Introspection.get_docstring

Term.Introspection.inspect

Term.Introspection.style_methods

Term.Introspection.get_docstring — Method

get_docstring(obj)

Extract and style an object's docstring.

Term.Introspection.inspect — Method

inspect(T::DataType; documentation::Bool=false, constructors::Bool=true, methods::Bool=true, supertypes::Bool=true)

Inspect a DataType to show info such as docstring, constructors and methods. Flags can be used to choose the level of detail in the information presented:

documentation: show docstring with termshow

constructors: show T constructors

methods: show methods using T in their signature

supertypes: show methods using T's supertypes in their signature

Term.Introspection.style_methods — Method

style_methods(methods::Union{Vector{Base.Method}, Base.MethodList}, tohighlight::AbstractString)

Create a Renderable with styled Method information for inspect(::DataType)