Layout

Term.layout.Padding

Term.layout.Padding

Term.layout.Spacer

Term.layout.hLine

Term.layout.hLine

Term.layout.hLine

Term.layout.hLine

Term.layout.hLine

Term.layout.vLine

Term.layout.vLine

Term.layout.vLine

Term.layout.hstack

Term.layout.hstack

Term.layout.vstack

Term.layout.vstack

Term.layout.Padding — Type

Padding

Stores empty string to pad a string to the desired with.

Term.layout.Padding — Method

Padding(text::AbstractString, target_width::Int, method::Symbol)::Padding

Create Padding for a string given a justification method ∈ (:left, :center, :right).

Term.layout.Spacer — Type

    Spacer

A box of empty text with given width and height.

Term.layout.hLine — Type

hLine

A 1-line renderable made of repeated character from a Box.

Term.layout.hLine — Method

hLine(text::AbstractString; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Construct an hLine as wide as the stdout with centered text.

Term.layout.hLine — Method

hLine(width::Number, text::String; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Creates an hLine object with texte centered horizontally.

Term.layout.hLine — Method

hLine(width::Number, style::Union{String, Nothing}; box::Symbol=:ROUNDED)

Create a styled hLine of given width.

Term.layout.hLine — Method

hLine(; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Construct an hLine as wide as the stdout

Term.layout.vLine — Type

vLine

A multi-line renderable with each line made of a | to create a vertical line

Term.layout.vLine — Method

vLine(height::Number, style::Union{String, Nothing}; box::Symbol=:ROUNDED)

Create a vLine given a height and, optionally, style information.

Term.layout.vLine — Method

vLine(; style::Union{String, Nothing}=nothing, box::Symbol=:ROUNDED)

Create a vLine as tall as the stdout console

Term.layout.hstack — Method

hstack(renderables...)

Horizonatlly stack a variable number of renderables

Term.layout.hstack — Method

hstack(r1::RenderablesUnion, r2::RenderablesUnion)

Horizontally stack two renderables to give a new renderable.

Term.layout.vstack — Method

vstack(renderables...)

Vertically stack a variable number of renderables

Term.layout.vstack — Method

vstack(r1::RenderablesUnion, r2::RenderablesUnion)

Vertically stack two renderables to give a new renderable.