Term Markdown

Term.Renderables.RenderableText

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.TermMarkdown.parse_md

Term.Tprint.tprint

Term.Tprint.tprintln

Term.Renderables.RenderableText — Method

RenderableText(md::Markdown.MD; width = console_width() - 2, kwargs...)

Create a RenderableText from a markdown string.

Term.TermMarkdown.parse_md — Function

parse_md

Parse a Markdown element (Paragraph, List...) into a string.

A width keyword argument can be used to control the width of the string representation and an inline boolean argument specifies when an element (e.g. a code snippet) is in-line within a larger element (e.g. a paragraph).

Term.TermMarkdown.parse_md — Method

function parse_md(ad::Markdown.Admonition; width = console_width(), kwargs...)::String

Parse adomitions and style them with colored Panel renderables.

Term.TermMarkdown.parse_md — Method

function parse_md(
    qt::Markdown.BlockQuote;
    width = default_width(),
    kwargs...,
)::String

Style a BlockQuote with a line and a quotation marker.

Term.TermMarkdown.parse_md — Method

function parse_md(
    code::Markdown.Code;
    width = console_width(),
    inline = false,
    kwargs...,
)::String

Parse a code snippet with syntax highlighting (for Julia code).

For non-inline code snippets the code is put in a panel with different background coloring to make it stand out.

Term.TermMarkdown.parse_md — Method

parse_md(note::Markdown.Footnote; width = console_width(), inline = false)

Style a footnote differently based on if they are a renference to it or its content.

Term.TermMarkdown.parse_md — Method

function parse_md(
    ltx::Markdown.LaTeX;
    inline = false,
    width = console_width(),
    kwargs...,
)::String

Parse a math element. Try to convert it to unicode characters when possible, otherwise output the latex string.

Term.TermMarkdown.parse_md — Method

function parse_md(
    list::Markdown.List;
    width = console_width(),
    inline = false,
    space = "",
)::String

Parse a list and all its elements, for both numbered and unnumbered lists.

Term.TermMarkdown.parse_md — Method

parse_md(text::Markdown.MD; kwargs...)::String

Parse an entier MD object by parsing its constituent elements and joining the resulting strings.

Term.TermMarkdown.parse_md — Method

parse_md(paragraph::Markdown.Paragraph; width = console_width(), kwargs...)::String

Parse each element in a paragraph

Term.TermMarkdown.parse_md — Method

function parse_md(tb::Markdown.Table; width = console_width())::String

Convert a markdown Table to a Table renderable.

Term.TermMarkdown.parse_md — Method

parse_md(header::Markdown.Header{l}; width = console_width(), kwargs...) where {l}

Parse Headers with different style based on the level

Term.Tprint.tprint — Method

tprint(md::Markdown.MD; kwargs...)

Print a parsed markdown string.

Term.Tprint.tprintln — Method

tprintln(md::Markdown.MD; kwargs...)

Print a parsed markdown string.