Measure

Term.measure.Measure

Term.measure.Measure

Base.:+

Term.measure.Measure — Type

Measure

Stores the size of a piece of renderable material

Term.measure.Measure — Method

Measure(str::String)

Constructs a measure object from a string

Base.:+ — Method

The sum of measures returns a measure with the highest value along each dimension