Panel

Term.panel.Panel

Term.panel.Panel

Term.panel.Panel

Term.panel.Panel

Term.panel.TextBox

Term.panel.TextBox

Term.panel.render

Term.panel.Panel — Type

Panel

Renderable with a panel surrounding some content: ╭──────────╮ │ my panel │ ╰──────────╯

Term.panel.Panel — Method

Panel(
    content::AbstractRenderable;
    width::Int = 88,
    height::Union{Nothing,Int} = nothing,
    fit::Bool=false,
    padding::Union{Padding, NTuple} = Padding(2, 2, 0, 0),
    kwargs...
)

Construct a Panel around of a AbstractRenderable

Term.panel.Panel — Method

Panel(renderables; kwargs...)

Panel constructor for creating a panel out of multiple renderables at once.

Term.panel.Panel — Method

Panel(; 
    fit::Symbol=:nofit,
    width::Int = 88,
    height::Union{Nothing,Int} = nothing, 
    kwargs...
)

Construct a Panel with no content

Term.panel.TextBox — Type

TextBox

Creates a Panel and fits input text to it. The pannel is hidden so that the result is just a text box.

Term.panel.TextBox — Method

TextBox(
    text::Union{Vector, AbstractString};
    width::Union{Nothing, Int}=nothing,
    title::Union{Nothing, String}=nothing,
    title_style::Union{String, Nothing}="default",
    title_justify::Symbol=:left,
    subtitle::Union{String, Nothing}=nothing,
    subtitle_style::Union{String, Nothing}="default",
    subtitle_justify::Symbol=:left,
    justify::Symbol=:left,
    fit::Symbol=:fit,
    )

Creates an hidden Panel with text in it.

If a width is passed, the input text is reshaped to have that size, unless fit=:truncate in which case it's cut to size. If no width is passed and fit=true the TextBox's size matches the size of the input text. Other arguments behave like Panel.

See also Panel.

Term.panel.render — Method

Panel(
        style::String,
        title_style::String,
        title_justify::Symbol,
        subtitle::String,
        subtitle_style::String,
        subtitle_justify::String,
        panel_measure::Measure,
        content_measure::Measure,
        Δw::Int,
        Δh::Int,
        padding::Padding
    )

Construct a Panel given all required info.