Panels

Term.Panels.Panel

Term.Panels.Panel

Term.Panels.Panel

Term.Panels.Panel

Term.Panels.Panel

Term.Panels.Panel

Term.Panels.render

Term.Panels.trim_renderable

Term.Panels.Panel — Type

Panel

Renderable with a panel surrounding some content:

    ╭──────────╮
    │ my panel │
    ╰──────────╯

Term.Panels.Panel — Method

Panel(
    content::Union{AbstractString,AbstractRenderable},
    ::Val{false},
    padding::Padding;
    height::Union{Nothing,Int} = nothing,
    width::Int = 76,
    kwargs...,
)

Construct a Panel fitting content to it.

Tip

Content that is too large to fit in the given width will be trimmed. To avoid trimming, set fit=true when calling panel.

Term.Panels.Panel — Method

Panel(
    content::Union{AbstractString,AbstractRenderable},
    ::Val{true},
    padding::Padding;
    height::Union{Nothing,Int} = nothing,
    width::Union{Nothing,Int} = nothing,
    trim::Bool = true,
    kwargs...,
)

Construct a Panel fitting the content's width.

Warning

If the content is larger than the console terminal's width, it will get trimmed to avoid overflow, unless trim=false is given.

Term.Panels.Panel — Method

Panel(
    content::Union{AbstractString,AbstractRenderable};
    fit::Bool = true,
    padding::Union{Padding,NTuple} = Padding(2, 2, 0, 0),
    kwargs...,
)

Construct a Panel around an AbstractRenderable or AbstractString.

This is the main Panel-creating function, it dispatches to other methods based on the value of fit to either fith the Panel to its content or vice versa.

kwargs can be used to set various aspects of the Panel's appearance like the presence and style of titles, box type etc... see render below.

Term.Panels.Panel — Method

Panel(renderables; kwargs...)

Panel constructor for creating a panel out of multiple renderables at once.

Term.Panels.Panel — Method

Panel(; 
    fit::Bool = false,
    height::Int = 2,
    width::Int = 76,
    padding::Union{Vector,Padding,NTuple} = Padding(0, 0, 0, 0),
    kwargs...,  
)

Construct a Panel with no content.

Examples

julia> Panel(height=5, width=10)
╭────────╮
│        │
│        │
│        │
╰────────╯

julia> Panel(height=3, width=5)
╭───╮
│   │
╰───╯

Term.Panels.render — Method

render(
    content;
    box::Symbol = :ROUNDED,
    style::String = "default",
    title::Union{String,Nothing} = nothing,
    title_style::Union{Nothing,String} = nothing,
    title_justify::Symbol = :left,
    subtitle::Union{String,Nothing} = nothing,
    subtitle_style::Union{Nothing,String} = nothing,
    subtitle_justify::Symbol = :left,
    justify::Symbol = :left,
    panel_measure::Measure,
    content_measure::Measure,
    Δw::Int,
    Δh::Int,
    padding::Padding,
)::Panel

Construct a Panel's content.

Term.Panels.trim_renderable — Method

trim_renderable(ren::Union{String, AbstractRenderable}, width::Int)

Trim a string or renderable to a max width.