ProgressBars

Term.progress.ProgressBar — Type

ProgressBar

Progress bar Type, stores information required to render a progress bar renderable.

Term.progress.TempSTDOUT — Type

TempSTDOUT

Stores information about a temporarily re-directed STDOUT

Term.progress.Track — Type

Track

Convenience iterator to add a ProgressBar to for loops.

for i in track(1:100)
    # ... do something
end

Adds a progress bar that updates for each iteration through the loop.

Base.iterate — Method

Update progress bar and continue iteration.

Base.iterate — Method

Start iteration. Crate Progress Bar

Term.progress.pbar_color — Method

pbar_color(pbar::ProgressBar)

Get the RGB color of of a progress bar's bar based on progress.

Term.progress.track — Method

Costructor for a Track object.

Term.progress.update — Method

update(pbar::ProgressBar)

Update progress bar info and display.