ProgressBars

Term.progress.ProgressBar — Type

ProgressBar

Progressbar Type, stores information required to render progress bar renderables for each ProgressJob assigned.

ProgressBar takes care of the work needed to setup/update progress bar(s) visualizations. Each individual bar corresponds to a (running) ProgressJob and ProgressJob itself is what actually creates the visuals. Most of the work done by ProgressBar is to handling terminal stuff, move cursor position, change scrolling regions, clear/update sections etc.

Arguments: - jobs: vector of ProgressJob to assign to the progress bar (more can added later) - width: width of the visualization, if expand=false - columns: which columns to show? Either a vector of column types or the name of a preset - column_kwargs: keyword arguments to pass to each column - transient: if true jobs disappear when done (and the whole pbar when all jobs are done) - colors: set of 3 RGBColor to change the color of the progress bar with progress - Δt: delay between refreshes of the visualization. - buff: IOBuffer used to render the progress bars - running: true if the progress bar is active - paused: false when the bar is running but briefly paused (e.g. to update jobs) - task: references a Task for updating the progress bar in parallel - renderstatus: a RenderStatus instance.

Term.progress.ProgressBar — Method

ProgressBar(;
    width::Int=88,
    columns::Union{Vector{DataType}, Symbol} = :default,
    columns_kwargs::Dict = Dict(),
    expand::Bool=false,
    transient::Bool = false,
    colors::Vector{RGBColor} = [
        RGBColor("(1, .05, .05)"),
        RGBColor("(.05, .05, 1)"),
        RGBColor("(.05, 1, .05)"),
    ],
    refresh_rate::Int=60,  # FPS of rendering
)

Create a ProgressBar instance.

Term.progress.ProgressJob — Type

ProgressJob

Single job whose progress we're tracking with a progress bar.

Each progress bar can have multiple jobs running at the same time, and more can be added/removed at any time. Each ProgressJob keeps track of the state (progress) of each of these jobs. ProgressJob is also rendered to create the visual display of the progress bar.

Arguments:

id specifies the job;s unique id

i: keeps track of the progress (number of completed steps)

N: total number of steps in the task at hand. Optional, set to nothing when not known

description: a bit of text describing what the job is

columns: set of AbstractColumn used to represent the job's information

columns_kwargs: additional information passed to each column when created (e.g. to set style)

width: width of the progress bar display

started: whether the job was started or not.

stopped: whether the job was stopped (finished job)

startime: time at which job was started

stoptime: time at which job was stopped

transient: if truee the job's visual display disappears when stopped=true

Term.progress.RenderStatus — Type

RenderStatus

Keep track of rendering information for a ProgressBar

Term.progress.addjob! — Method

addjob!(
        pbar::ProgressBar;
        description::String="Running...",
        N::Union{Int, Nothing}=nothing,
        start::Bool=true,
        transient::Bool=false,
        id=nothing
    )::ProgressJob

Add a new ProgressJob to a running ProgressBar

See also: removejob!, getjob

Term.progress.getjob — Method

getjob(pbar::ProgressBar, id)

Get a ProgressBar's ProgressJob by id.

Term.progress.jobcolor — Method

jobcolor(job::ProgressJob)

Get the RGB color of of a progress bar's bar based on progress.

Term.progress.removejob! — Method

removejob!(pbar::ProgressBar, job::ProgressJob)

Remove a ProgressJob from a ProgressBar.

See also: addjob!, getjob

Term.progress.render — Method

render(job::ProgressJob, pbar::ProgressBar)::String

Render a ProgressBar.

When a progress bar is first rendered, this function uses ANSI codes to change the scrolling region of the terminal window to create a space at the bottom where the bar's visuals can be displayed. This allows for thext printed to stdout to still be visualized. On subsequent calls, this function ensures that the height of the reserved space matches the number of running jobs.

All fo this requires a bit of careful work in moving the cursor around and doing ANSI magic.

Term.progress.render — Method

render(job::ProgressJob, pbar::ProgressBar)::String

Render a ProgressJob

Term.progress.render — Method

render(job::ProgressJob, pbar::ProgressBar)::String

Render a ProgressJob

Term.progress.start! — Method

start!(pbar::ProgressBar)

Start a ProgressBar and run a Task to update its visuals.

Starts a parallel Task for updating the ProgressBar visualization while other code runs. The task is stopped when the progress bar is set to have running=false

See also stop!

Term.progress.start! — Method

start!(job::ProgressJob)

Start a newly created ProgressJob.

When starting a job, take care of creating instance of AbstractColumns to display the job's progress. If N is nothing, remove any ProgressColumn

Term.progress.stop! — Method

stop!(pbar::ProgressBar)

Stop a running ProgressBar.

Stops the Task updating the progress bar visuals too, and if the progress bar was transient it clears up the visuals.

Term.progress.stop! — Method

stop!(job::ProgressJob)

Stop a running job.

Term.progress.update! — Method

update!(job::ProgressJob; i = nothing)

Update a job's progress i by setting its value or adding +1.

Term.progress.with — Method

with(expr, pbar::ProgressBar)

Wrap an expression to run code in the context of a progress bar.

Ensures that the progress bar is correctly created and distroyed even if code within the loop whose progress we're monitoring causes and error. Since render changes the scrollregion in the terminal, we need to make sure that we can restore things no matter what.

Examples

pbar = ProgressBar()
with(pbar) do
    job = addjob!(pbar; description="Running")
    for i in 1:500
        update!(job)
        sleep(.0025)
    end
end

Term.progress.@track — Macro

@track(ex)

Macro to wrap a loop in a progress bar.

Like with it wraps the expression code in a try/finally statement to ensure that the progress bar is correctly stopped in any situation.

Examples

@track for i in 1:10
    sleep(0.1)
end