Renderables

Term.Renderables.AbstractRenderable

Term.Renderables.Renderable

Term.Renderables.RenderableText

Term.Renderables.RenderableText

Base.print

Base.show

Base.show

Base.string

Term.Renderables.trim_renderable

Term.Renderables.AbstractRenderable — Type

AbstractRenderable

Term.Renderables.Renderable — Method

Renderable(
    str::String; width::Union{Nothing,Int} = nothing
)

Convenience method to construct a RenderableText

Term.Renderables.RenderableText — Method

RenderableText(text::String; width::Union{Nothing, Int, Symbol}=nothing)

Construct a RenderableText out of a string.

The text is resized to fit the given width. Optionally justify can be used to set the text justification style ∈ (:left, :center, :right, :justify).

Term.Renderables.RenderableText — Method

RenderableText(rt::RenderableText; width::Union{Nothing,Int} = nothing)

Construct a RenderableText by possibly re-shaping a RenderableText

Base.print — Method

print(io::IO, renderable::AbstractRenderable)

Print a renderable to an IO

Base.show — Method

show(io::IO, mime::MIME"text/plain", renderable::AbstractRenderable)

Show a renderable and some information about its shape.

Base.show — Method

show(io::IO, renderable::AbstractRenderable)

Show a renderable.

Base.string — Method

Base.string(r::AbstractRenderable)::String

Creates a string representation of a renderable

Term.Renderables.trim_renderable — Method

trim_renderable(ren::Union{String, AbstractRenderable}, width::Int)

Trim a string or renderable to a max width.