Style

Term.style.MarkupStyle

Term.style.MarkupStyle

Term.style.apply_style

Term.style.get_style_codes

Term.style.is_mode

Term.style.MarkupStyle — Type

MarkupStyle

Holds information about the style specification set out by a MarkupTag.

Term.style.MarkupStyle — Method

MarkupStyle(tag::MarkupTag)

Builds a MarkupStyle definition from a MarkupTag.

Term.style.apply_style — Method

apply_style(text)

Apply style to a piece of text.

Extract markup style information and insert the appropriate ANSI codes to style a string.

Term.style.get_style_codes — Method

get_style_codes(style::MarkupStyle)

Get ANSICodes corresponding to a MarkupStyle.

Term.style.is_mode — Method

Check if a string is a mode name