Term

Term.OPEN_TAG_REGEX

Term.CodeTheme

Term.Theme

Term.measure.Measure

Term.measure.Measure

Highlights.Format.render

Term.chars

Term.cleantext

Term.do_by_line

Term.escape_brackets

Term.expr2string

Term.fillin

Term.get_ANSI_codes

Term.get_file_format

Term.get_last_ANSI_code

Term.get_lr_widths

Term.has_ansi

Term.has_markup

Term.highlight

Term.highlight

Term.highlight_syntax

Term.join_lines

Term.load_code_and_highlight

Term.loop_last

Term.nospaces

Term.or

Term.read_file_lines

Term.remove_ansi

Term.remove_brackets

Term.remove_markup

Term.replace_ansi

Term.replace_multi

Term.replace_text

Term.replace_text

Term.reshape_text

Term.split_lines

Term.split_lines

Term.textlen

Term.truncate

Term.tview

Term.unescape_brackets

Term.unspace_commas

Term.@make_color_macro

Term.@make_mode_macro

Term.@style

Term.OPEN_TAG_REGEX — Constant

This regex uses lookahead and lookbehind to exclude [[ at the beginning of a tag, with this: (?<![)[(?![)

Term.CodeTheme — Type

Custom hilighting theme for Highlighters.jl https://juliadocs.github.io/Highlights.jl/stable/man/theme/

Term.Theme — Type

Theme

Stores colors for different semantically relevant items, used to style outputs to terminal.

Term.measure.Measure — Method

Measure(seg::Segment)

gives the measure of a segment

Term.measure.Measure — Method

Measure(segments::Vector{Segment})

gives the measure of a vector of segments

Highlights.Format.render — Method

Format.render(io::IO, ::MIME"text/ansi", tokens::Format.TokenIterator)

custom ANSI lexer for Highlighters.jl

Term.chars — Method

Split a string into a vector of Chars.

Term.cleantext — Method

cleantext(str::AbstractString)

Remove all style information from a string.

Term.do_by_line — Method

do_by_line(fn::Function, text::String)

Apply fn to each line in the text.

The function fn should accept a single ::String argument.

Term.escape_brackets — Method

remove_ansi(str)::String

Replace each squared bracket with a double copy of itself

Term.expr2string — Method

Get a clean string representation of an expression

Term.fillin — Method

fillin(text::String)::String

Ensure that each line in a multi-line text has the same width.

Term.get_ANSI_codes — Method

get_ANSI_codes(text)::String

Returns a string with all ANSI codes in the input.

Term.get_file_format — Method

getfileformat(nbytes; suffix="B")

Return a string with formatted file size.

Term.get_last_ANSI_code — Method

get_last_ANSI_code(text)::String

Get the last ANSI code in a sting, returns "" if no ANSI code found.

Term.get_lr_widths — Method

getlrwidths(width::Int)

To split something with width in 2, get the lengths of the left/right widths.

When width is even that's easy, when it's odd we need to be careful.

Term.has_ansi — Method

has_ansi(text::String)

Returns true if text includes a MarkupTag

Term.has_markup — Method

has_markup(text::String)

Returns true if text includes a MarkupTag

Term.highlight — Method

highlight(text::AbstractString, theme::Theme, like::Symbol)

Hilights an entire text as if it was a type of semantically relevant text of type :like.

Term.highlight — Method

highlight(text::AbstractString, theme::Theme)

Highlighs a text introducing markup to style semantically relevant segments, colors specified by a theme object

Term.highlight_syntax — Method

highlight_syntax(code::AbstractString; style::Bool=true)

Highlight Julia code syntax in a string.

Term.join_lines — Method

join_lines(lines)

Merge a vector of strings in a single string.

Term.load_code_and_highlight — Method

load_code_and_highlight(path::AbstractString, lineno::Int; δ::Int=3, width::INt=120)

Load a file, get the code and format it. Return styled text

Term.loop_last — Method

loop_last(v)

Returns an iterable yielding tuples (is_last, value).

Term.nospaces — Method

nospaces(text::AbstractString)

Remove all spaces from a string.

Term.or — Method

or(x, y)

Return y if x is Nothing

Term.read_file_lines — Method

read_file_lines(path::String, start::Int, stop::Int)

Read a file and select only lines in range start -> stop.

Returns a vector of tuples with the line number and line content.

Term.remove_ansi — Method

remove_ansi(input_text::AbstractString)::AbstractString

Remove all ANSI tags from a string of text

Term.remove_brackets — Method

remove_brackets(text::AbstractString)

Remove all () brackets from a string.

Term.remove_markup — Method

remove_markup(input_text::AbstractString)::AbstractString

Remove all markup tags from a string of text.

Term.replace_ansi — Method

replace_ansi(input_text)

Replace ANSI tags with ¦.

The number of '¦' matches the length of the ANSI tags. Used when we want to hide ANSI tags but keep the string length intact.

Term.replace_multi — Method

multiple strings replacement, for multiple on Julia version

Term.replace_text — Function

replace_text(text::AbstractString, start::Int, stop::Int, char::Char='_')

Replace a section of a text between start and stop with another string composed of repeats of a given character char.

Term.replace_text — Method

replace_text(text::AbstractString, start::Int, stop::Int, replace::AbstractString)

Replace a section of a text between start and stop with replace.

Term.reshape_text — Method

reshape_text(text, width::Int)

Reshape a string to have max width when printed out.

Cut the string into multiple lines so that each line has at most width when printed to terminal: ignoring ANSI style but taking into account characters with widths > 1. Text is preferentially split at a space when possible to improve readability.

Term.split_lines — Method

split_lines(renderable)

Split a renderable's text.

Term.split_lines — Method

split_lines(text::AbstractString)

Split a string into its composing lines.

Term.textlen — Method

textlen(x::AbstractString)

Get length of text after all style information is removed.

Term.truncate — Method

truncate(text::AbstractString, width::Int)

Shorten a string of text to a target width

Term.tview — Method

tview(text, start::Int, stop::Int)

Get a view object with appropriate indices

Term.unescape_brackets — Method

unescape_brackets(text)::String

Replece every double squared parenthesis with a single copy of itself

Term.unspace_commas — Method

unspace_commas(text::AbstractString)

Remove spaces after commas.

Term.@make_color_macro — Macro

Macro to create macros such as @green which colors text accordingly

Term.@make_mode_macro — Macro

Macro to create macros such as @underline which styles text accordingly.

Term.@style — Macro

@style "text" style1 style2...

Applies a sequence of styles to a piece of text, such that

println(@style "my text" bold green underline)

will print my text as bold, green and underlined